# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API",
        "description": "API documentation for the Pet Shop application",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/main/blog/{uuid}": {
            "get": {
                "tags": [
                    "Main"
                ],
                "summary": "Get blog post information",
                "description": "Returns blog post data",
                "operationId": "showBlog",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Blog post uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Blog post not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}